requirements.txt
flask
gunicorn
opencv-python
cmake
dlib
face_recognition
pillow
mysql
mysql-connector-python